engine/requirements.txt
# Astra Neural Engine - Python Dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
langgraph>=0.0.26
langchain>=0.1.0
langchain-core>=0.1.17
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
python-dotenv>=1.0.0
agentops>=0.3.0
